🔧 feat(prisma): change `Period` field type to `Boolean`
The `Period` field in both the `proctor_in_room` and `exam_mission` models has been changed from an `Int` type with a default value of 0 to a `Boolean` type with a default value of `false`. This change simplifies the data model and better represents the boolean nature of the "period" concept.