🎨 feat(exam-mission): update `Period` property type
The `Period` property in the `CreateExamMissionDto` class has been updated from a `number` type to a `Boolean` type. This change was made to better represent the intended usage of the `Period` field, which likely indicates whether an exam mission has a specific duration or not, rather than a numerical value.