🎨 feat(proctor): update proctor findAllBySchoolId endpoint
The changes in this commit update the `findAllBySchoolId` endpoint in the ProctorController.
The previous implementation used a `school_id` parameter, but this has been changed to
retrieve the `Schools_ID` from the request headers, which is a more secure and
convenient approach, as the user's school ID is already available in the authenticated
request.