🎨 feat(ExamMissionService): Add createdBy parameter to create method
The changes in this commit add a new `createdBy` parameter to the `create` method of the `ExamMissionService`. This allows the service to associate the created exam mission with the user who created it. Additionally, the `data` object passed to the `prisma.exam_mission.create` method is now spread with the `createdBy` value.