✨feat: Add `Created_By` field to school class creation
The changes made in this commit add the `Created_By` field to the `create` method of the `SchoolClassesService`. This allows the creation of a school class to be associated with the user who created it. This is an important feature to track the ownership and accountability of school class creation.