🎨 feat(school-classes): Add user ID to create school class
This change adds the user's ID to the `create` method in the `SchoolClassesController`. This allows the `SchoolClassesService` to associate the created school class with the user who created it.

The changes include:

- Adding `@Req()` to the `create` method to get the user's ID from the request headers
- Passing the user's ID to the `schoolClassesService.create()` method

This feature will help to track the ownership and responsible user for each school class created in the system.