🎨 feat(students): Add user ID to create and create many students
This commit adds the user ID to the create and create many students endpoints
in the student controller. This is done to ensure that the created students
are associated with the user who created them. This change is important to
maintain the integrity of the student data and to provide a way to track the
user responsible for creating each student.