✨ feat(school-classes): add `updatedBy` parameter to update method
The `update` method in the `SchoolClassesService` has been modified to accept an additional `updatedBy` parameter. This change allows the caller to provide the ID of the user who performed the update operation. The updated data object now includes the `Updated_By` field, which will be stored in the database along with the other updated fields.