🚀 feat(prisma/schema): Update `Updated_By` fields to use `Int` type
The changes in this commit update the `Updated_By` fields in several Prisma models to use the `Int` data type instead of `String`. This change is made to ensure consistency and better data integrity within the schema.

The affected models are:
- `control_mission`
- `schools`
- `users`
- `exam_room`

This change will allow for more precise and accurate tracking of the users responsible for updating these records, as an integer ID can provide a stronger link to the corresponding user entity.