🔒 feat(auth): remove password from auth response
The changes made in this commit aim to remove the password field from the
response of the auth service. This is done to improve security and
prevent the password from being exposed to the client. The password field
is set to undefined before returning the user object in the various auth
methods.