✨ feat: Add endpoint to mark student as attended
The changes include adding a new endpoint `POST /student-barcodes/mark-attended/:barcode` to the `StudentBarcodesController`. This endpoint allows a student to mark themselves as attended by providing their barcode. The implementation of this feature is delegated to the `studentBarcodesService.markStudentAttended()` method.

This new functionality is intended to provide a way for students to easily check-in to an exam or mission by scanning their barcode, improving the overall experience.