🎨 feat(StudentService): unmark student as cheating
Adds a new method `unmarkAsCheating` to the `StudentService` that allows a
proctor to unmark a student as cheating. The method checks if the proctor
belongs to the same division as the student before updating the `isCheating`
flag in the database. If the proctor does not belong to the same division, a
`403 Forbidden` error is thrown.