🎨 feat(SchoolClassesController): Add Control Officer role for CRUD operations
Adds the Control Officer role to the `@Roles` decorator for the following methods in the `SchoolClassesController`:

- `create()`
- `remove()`
- `update()`
- `findBySchool()`

This change allows Control Officers, in addition to Super Admins, to perform CRUD operations on school classes. This provides more granular access control and allows Control Officers to manage school classes as part of their responsibilities.