🚫 Disable superadmin authorization for proctor management
The changes in this commit aim to remove the `@Roles(Role.SuperAdmin)` authorization check from various endpoints in the `ProctorController`. This change is made to allow users with different roles, other than the superadmin, to access and manage proctor-related functionality.

By disabling the superadmin-only authorization, the application becomes more accessible and flexible, allowing a wider range of users to perform proctor-related operations, such as creating, updating, deleting, and unassigning proctors from exam rooms.