🔖 feat(schema): add unique constraint and index for school class and cohort
Added a unique constraint on the `Schools_ID` and `Name` fields of the `school_class` model to ensure that each school class name is unique within a school.
Added an index on the `Name` and `School_Type_ID` fields of the `cohort` model to improve query performance when searching for cohorts by name and school type.