✨ feat(package): update Prisma to version 5.21.1
This change updates the Prisma dependency from version 5.20.0 to 5.21.1 in both the `package.json` and `package-lock.json` files. The primary goal of this update is to take advantage of the latest features and bug fixes provided by the Prisma library, which is a crucial component of the application's data management and interaction with the database.